+++ src/EmbeddedLevelNameEncoding.C
@@ -35,7 +35,7 @@ char* EmbeddedLevelNameEncoding::nameById(uint64 id) {
 	} else {
 		returnedName[0] = 'S';
 	}
-	for(uint64 i=1;i<nameSize-1;i++) {
+	for(int i=1;i<nameSize-1;i++) {
 		int c = '0' + (int) ((id >> (62 - 2*i)) & (uint32) 3);
 		returnedName[i] = (char) c;
 	}
@@ -259,7 +259,7 @@ void EmbeddedLevelNameEncoding::setIdFromSciDBLeftJustifiedFormat( int64 id_scid
  * @param level
  */
 EmbeddedLevelNameEncoding EmbeddedLevelNameEncoding::atLevel(uint64 level, bool keepAllBits ) {
-	uint oldLevel = this->getLevel();
+	uint32 oldLevel = this->getLevel();
 	uint64 id_NoLevel = this->maskOffLevel();
 	uint64 keepBits = one << 1; // Position 63
 	keepBits++; // Position 62
